eina: improve Eina_Clist support for 64bits system.
This warning are triggered by a cross compilation as previously stated.

+++ src/lib/eina/eina_clist.h
@@ -420,7 +420,7 @@ static inline void eina_clist_move_head(Eina_Clist *dst, Eina_Clist *src);
  */
 #undef EINA_CLIST_ENTRY
 #define EINA_CLIST_ENTRY(elem, type, field) \
-    ((type *)((char *)(elem) - (unsigned long)(&((type *)0)->field)))
+    ((type *)((char *)(elem) - (uintptr_t)(&((type *)0)->field)))
 
 #include "eina_inline_clist.x"
